Fix running buck in virtualenv (#14)
Buck was broken when run in a virtualenv inside the buck project, since a core python file becomes
subject to `allow_unsafe_import(false)`. This change detects whether the file we're parsing is inside
a virtualenv, and doesn't block imports if we are.

See facebook#2162 for a longer bug description and investigation.