Standardize the output of JUnit reports (#24)
Between testNG and JUnit, there were different output locations. This
makes it difficult for test reporters to intermingle: if a repo-wide
test run contains both TestNG and JUnit tests, then test reporters will
pick up only TestNG.

This is a short-term fix to get TestNG and JUnit reporting in the same
position. In the end, buck is deprecated, so we won't really invest
further.